+++ packages/runtime/src/interpreter/opcode-list.ts
@@ -24,17 +24,27 @@ import { RUNTIME_ERRORS } from "../errors/errors-list";
 import { RuntimeError } from "../errors/runtime-errors";
 import { compareArray } from "../lib/compare";
 import {
+	AccountParamGetField,
 	AcctParamQueryFields,
+	AcctParamQueryFieldsIndex,
 	ALGORAND_MAX_LOGS_COUNT,
 	ALGORAND_MAX_LOGS_LENGTH,
 	AppParamDefined,
+	AppParamDefinedIndex,
+	AppParamField,
 	AssetHoldingField,
 	AssetParamGetField,
-	AppParamField,
 	AssetParamMap,
+	AssetParamMapIndex,
+	Base64Encoding,
+	blockFieldIndex,
 	blockFieldTypes,
+	CurveTypeArgument,
 	CurveTypeEnum,
+	CurveTypeIndex,
+	GlobalField,
 	GlobalFields,
+	GlobalFieldsIndex,
 	ITxArrFields,
 	json_refTypes,
 	MathOp,
@@ -46,17 +56,16 @@ import {
 	MAX_UINT64,
 	MAX_UINT128,
 	MaxTEALVersion,
+	NumIndex,
 	OpGasCost,
 	proofLength,
 	publicKeyLength,
 	TransactionTypeEnum,
 	TxArrFields,
+	TxFieldEnum,
+	TxnFieldsIndex,
 	vrfVerifyFieldTypes,
 	ZERO_ADDRESS,
-	TxFieldEnum,
-	Base64Encoding,
-	GlobalField,
-	AccountParamGetField
 } from "../lib/constants";
 import { addInnerTransaction, calculateInnerTxCredit, setInnerTxField } from "../lib/itxn";
 import { bigintSqrt } from "../lib/math";
@@ -1444,8 +1453,9 @@ export class Txn extends Op {
 		this.line = line;
 		this.idx = undefined;
 
-		this.assertTxFieldDefined(args[0], interpreter.tealVersion, line);
-		if (TxArrFields[interpreter.tealVersion].has(args[0])) {
+		const argument = this.assertFieldIndex(args[0], TxnFieldsIndex);
+		this.assertTxFieldDefined(argument, interpreter.tealVersion, line);
+		if (TxArrFields[interpreter.tealVersion].has(argument)) {
 			// eg. txn Accounts 1
 			assertLen(args.length, 2, line);
 			assertOnlyDigits(args[1], line);
@@ -1453,9 +1463,8 @@ export class Txn extends Op {
 		} else {
 			assertLen(args.length, 1, line);
 		}
-		this.assertTxFieldDefined(args[0], interpreter.tealVersion, line);
 
-		this.field = args[0]; // field
+		this.field = argument; // field
 		this.interpreter = interpreter;
 	}
 
@@ -1516,9 +1525,11 @@ export class Gtxn extends Op {
 			assertLen(args.length, 2, line);
 		}
 		assertOnlyDigits(args[0], line);
-		this.assertTxFieldDefined(args[1], interpreter.tealVersion, line);
+
+		const argument = this.assertFieldIndex(args[1], TxnFieldsIndex);
+		this.assertTxFieldDefined(argument, interpreter.tealVersion, line);
 		this.txIdx = Number(args[0]); // transaction group index
-		this.field = args[1]; // field
+		this.field = argument; // field
 		this.groupTxn = interpreter.runtime.ctx.gtxs;
 		this.interpreter = interpreter;
 	}
@@ -1565,9 +1576,10 @@ export class Txna extends Op {
 		super();
 		this.line = line;
 		assertOnlyDigits(args[1], line);
-		this.assertTxArrFieldDefined(args[0], interpreter.tealVersion, line);
+		const argument = this.assertFieldIndex(args[0], TxnFieldsIndex);
+		this.assertTxArrFieldDefined(argument, interpreter.tealVersion, line);
 
-		this.field = args[0]; // field
+		this.field = argument; // field
 		this.fieldIdx = Number(args[1]);
 		this.interpreter = interpreter;
 	}
@@ -1620,10 +1632,12 @@ export class Gtxna extends Op {
 		assertLen(args.length, 3, line);
 		assertOnlyDigits(args[0], line);
 		assertOnlyDigits(args[2], line);
-		this.assertTxArrFieldDefined(args[1], interpreter.tealVersion, line);
+
+		const argument = this.assertFieldIndex(args[1], TxnFieldsIndex);
+		this.assertTxArrFieldDefined(argument, interpreter.tealVersion, line);
 
 		this.txIdx = Number(args[0]); // transaction group index
-		this.field = args[1]; // field
+		this.field = argument; // field
 		this.fieldIdx = Number(args[2]); // transaction field array index
 		this.groupTxn = interpreter.runtime.ctx.gtxs;
 		this.interpreter = interpreter;
@@ -1849,9 +1863,10 @@ export class Global extends Op {
 	constructor(args: string[], line: number, interpreter: Interpreter) {
 		super();
 		assertLen(args.length, 1, line);
-		this.assertGlobalDefined(args[0], interpreter.tealVersion, line);
+		const argument = this.assertFieldIndex(args[0], GlobalFieldsIndex);
+		this.assertGlobalDefined(argument, interpreter.tealVersion, line);
 
-		this.field = args[0]; // global field
+		this.field = argument; // global field
 		this.interpreter = interpreter;
 		this.line = line;
 	}
@@ -2340,9 +2355,11 @@ export class GetAssetHolding extends Op {
 		}
 		let value: StackElem;
 		switch (this.field) {
+			case NumIndex.index_0:
 			case AssetHoldingField.AssetBalance:
 				value = BigInt(assetInfo.amount);
 				break;
+			case NumIndex.index_1:
 			case AssetHoldingField.AssetFrozen:
 				value = assetInfo["is-frozen"] ? 1n : 0n;
 				break;
@@ -2377,7 +2394,8 @@ export class GetAssetDef extends Op {
 		this.line = line;
 		this.interpreter = interpreter;
 		assertLen(args.length, 1, line);
-		if (AssetParamMap[interpreter.tealVersion][args[0]] === undefined) {
+		const argument = this.assertFieldIndex(args[0], AssetParamMapIndex);
+		if (AssetParamMap[interpreter.tealVersion][argument] === undefined) {
 			throw new RuntimeError(RUNTIME_ERRORS.TEAL.UNKNOWN_ASSET_FIELD, {
 				field: args[0],
 				line: line,
@@ -2385,7 +2403,7 @@ export class GetAssetDef extends Op {
 			});
 		}
 
-		this.field = args[0];
+		this.field = argument;
 	}
 
 	execute(stack: TEALStack): number {
@@ -2421,7 +2439,9 @@ export class GetAssetDef extends Op {
 					break;
 				default:
 					def = AssetDefinition[s] as string;
-					value = isValidAddress(def) ? decodeAddress(def).publicKey : parsing.stringToBytes(def);
+					value = isValidAddress(def)
+						? decodeAddress(def).publicKey
+						: parsing.stringToBytes(def);
 					break;
 			}
 
@@ -3900,17 +3920,17 @@ export class EcdsaVerify extends Op {
 		super();
 		this.line = line;
 		assertLen(args.length, 1, line);
-		this.curveIndex = Number(args[0]);
-		if (this.curveIndex === 0) {
-			this.curveType = CurveTypeEnum.secp256k1;
-		} else if (this.curveIndex === 1) {
-			this.curveType = CurveTypeEnum.secp256r1;
-		} else {
+		const argument = this.assertFieldIndex(args[0], CurveTypeIndex);
+		if (argument == CurveTypeArgument.secp256k1) this.curveIndex = 0;
+		else if (argument == CurveTypeArgument.secp256r1) this.curveIndex = 1;
+		else {
 			throw new RuntimeError(RUNTIME_ERRORS.TEAL.CURVE_NOT_SUPPORTED, {
 				line: this.line,
-				index: this.curveIndex,
+				curve: argument,
 			});
 		}
+
+		this.curveType = this.curveIndex === 0 ? CurveTypeEnum.secp256k1 : CurveTypeEnum.secp256r1;
 	}
 
 	computeCost(): number {
@@ -3960,17 +3980,16 @@ export class EcdsaPkDecompress extends Op {
 		super();
 		this.line = line;
 		assertLen(args.length, 1, line);
-		this.curveIndex = Number(args[0]);
-		if (this.curveIndex === 0) {
-			this.curveType = CurveTypeEnum.secp256k1;
-		} else if (this.curveIndex === 1) {
-			this.curveType = CurveTypeEnum.secp256r1;
-		} else {
+		const argument = this.assertFieldIndex(args[0], CurveTypeIndex);
+		if (argument == CurveTypeArgument.secp256k1) this.curveIndex = 0;
+		else if (argument == CurveTypeArgument.secp256r1) this.curveIndex = 1;
+		else {
 			throw new RuntimeError(RUNTIME_ERRORS.TEAL.CURVE_NOT_SUPPORTED, {
 				line: this.line,
-				index: this.curveIndex,
+				curve: argument,
 			});
 		}
+		this.curveType = this.curveIndex === 0 ? CurveTypeEnum.secp256k1 : CurveTypeEnum.secp256r1;
 	}
 
 	computeCost(): number {
@@ -4002,6 +4021,7 @@ export class EcdsaPkDecompress extends Op {
 export class EcdsaPkRecover extends Op {
 	readonly line: number;
 	readonly curveIndex: number;
+	readonly curveType: string;
 
 	/**
 	 * Asserts 1 arguments are passed.
@@ -4012,7 +4032,16 @@ export class EcdsaPkRecover extends Op {
 		super();
 		this.line = line;
 		assertLen(args.length, 1, line);
-		this.curveIndex = Number(args[0]);
+		const argument = this.assertFieldIndex(args[0], CurveTypeIndex);
+		if (argument == CurveTypeArgument.secp256k1) this.curveIndex = 0;
+		else if (argument == CurveTypeArgument.secp256r1) this.curveIndex = 1;
+		else {
+			throw new RuntimeError(RUNTIME_ERRORS.TEAL.CURVE_NOT_SUPPORTED, {
+				line: this.line,
+				curve: argument,
+			});
+		}
+		this.curveType = this.curveIndex === 0 ? CurveTypeEnum.secp256k1 : CurveTypeEnum.secp256r1;
 	}
 
 	computeCost(): number {
@@ -4031,14 +4060,7 @@ export class EcdsaPkRecover extends Op {
 		const recoverId = this.assertBigInt(stack.pop(), this.line);
 		const data = this.assertBytes(stack.pop(), this.line);
 
-		if (this.curveIndex !== 0) {
-			throw new RuntimeError(RUNTIME_ERRORS.TEAL.CURVE_NOT_SUPPORTED, {
-				line: this.line,
-				index: this.curveIndex,
-			});
-		}
-
-		const ec = new EC(CurveTypeEnum.secp256k1);
+		const ec = new EC(this.curveType);
 		const signature = { r: signatureC, s: signatureD };
 		const pubKey = ec.recoverPubKey(data, signature, Number(recoverId));
 		const x = pubKey.getX();
@@ -4244,9 +4266,10 @@ export class ITxnField extends Op {
 		super();
 		this.line = line;
 
-		this.assertTxFieldDefined(args[0], interpreter.tealVersion, line);
+		const argument = this.assertFieldIndex(args[0], TxnFieldsIndex);
+		this.assertTxFieldDefined(argument, interpreter.tealVersion, line);
 		assertLen(args.length, 1, line);
-		this.field = args[0]; // field
+		this.field = argument; // field
 		this.interpreter = interpreter;
 	}
 
@@ -4369,10 +4392,10 @@ export class ITxnSubmit extends Op {
 			const signedTransactions: algosdk.SignedTransaction[] = execParams.map((txnParam) =>
 				types.isExecParams(txnParam)
 					? {
-						sig: Buffer.alloc(5),
-						sgnr: Buffer.from(algosdk.decodeAddress(contractAddress).publicKey),
-						txn: webTx.mkTransaction(txnParam, mockSuggestedParams(txnParam.payFlags, 1)),
-					}
+							sig: Buffer.alloc(5),
+							sgnr: Buffer.from(algosdk.decodeAddress(contractAddress).publicKey),
+							txn: webTx.mkTransaction(txnParam, mockSuggestedParams(txnParam.payFlags, 1)),
+					  }
 					: txnParam
 			);
 			this.interpreter.runtime.ctx.processTransactions(signedTransactions);
@@ -4416,10 +4439,12 @@ export class ITxn extends Op {
 		super();
 		this.line = line;
 		this.idx = -1;
-		this.assertITxFieldDefined(args[0], interpreter.tealVersion, line);
+
+		const argument = this.assertFieldIndex(args[0], TxnFieldsIndex);
+		this.assertITxFieldDefined(argument, interpreter.tealVersion, line);
 		if (
-			TxArrFields[interpreter.tealVersion].has(args[0]) ||
-			ITxArrFields[interpreter.tealVersion].has(args[0])
+			TxArrFields[interpreter.tealVersion].has(argument) ||
+			ITxArrFields[interpreter.tealVersion].has(argument)
 		) {
 			// eg. itxn Accounts 1
 			assertLen(args.length, 2, line);
@@ -4428,7 +4453,7 @@ export class ITxn extends Op {
 		} else {
 			assertLen(args.length, 1, line);
 		}
-		this.field = args[0]; // field
+		this.field = argument; // field
 		this.interpreter = interpreter;
 	}
 
@@ -4458,9 +4483,10 @@ export class ITxna extends Op {
 		this.line = line;
 		assertLen(args.length, 2, line);
 		assertOnlyDigits(args[1], line);
-		this.assertITxArrFieldDefined(args[0], interpreter.tealVersion, line);
+		const argument = this.assertFieldIndex(args[0], TxnFieldsIndex);
+		this.assertITxArrFieldDefined(argument, interpreter.tealVersion, line);
 
-		this.field = args[0]; // field
+		this.field = argument; // field
 		this.idx = Number(args[1]);
 		this.interpreter = interpreter;
 	}
@@ -4728,8 +4754,8 @@ export class AppParamsGet extends Op {
 		this.line = line;
 		this.interpreter = interpreter;
 		assertLen(args.length, 1, line);
-
-		if (!AppParamDefined[interpreter.tealVersion].has(args[0])) {
+		const argument = this.assertFieldIndex(args[0], AppParamDefinedIndex);
+		if (!AppParamDefined[interpreter.tealVersion].has(argument)) {
 			throw new RuntimeError(RUNTIME_ERRORS.TEAL.UNKNOWN_APP_FIELD, {
 				field: args[0],
 				line: line,
@@ -4737,7 +4763,7 @@ export class AppParamsGet extends Op {
 			});
 		}
 
-		this.field = args[0];
+		this.field = argument;
 	}
 
 	execute(stack: Stack<StackElem>): number {
@@ -4802,10 +4828,10 @@ export class AcctParamsGet extends Op {
 		this.line = line;
 		this.interpreter = interpreter;
 		assertLen(args.length, 1, line);
-
+		const argument = this.assertFieldIndex(args[0], AcctParamQueryFieldsIndex);
 		if (
-			!AcctParamQueryFields[args[0]] ||
-			AcctParamQueryFields[args[0]].version > interpreter.tealVersion
+			!AcctParamQueryFields[argument] ||
+			AcctParamQueryFields[argument].version > interpreter.tealVersion
 		) {
 			throw new RuntimeError(RUNTIME_ERRORS.TEAL.UNKNOWN_ACCT_FIELD, {
 				field: args[0],
@@ -4814,7 +4840,7 @@ export class AcctParamsGet extends Op {
 			});
 		}
 
-		this.field = args[0];
+		this.field = argument;
 	}
 
 	execute(stack: Stack<StackElem>): number {
@@ -4992,10 +5018,12 @@ export class Base64Decode extends Op {
 		assertLen(args.length, 1, line);
 		const argument = args[0];
 		switch (argument) {
+			case NumIndex.index_0:
 			case Base64Encoding.URLEncoding: {
 				this.encoding = "base64url";
 				break;
 			}
+			case NumIndex.index_1:
 			case Base64Encoding.StdEncoding: {
 				this.encoding = "base64";
 				break;
@@ -5137,14 +5165,17 @@ export class Json_ref extends Op {
 		assertLen(args.length, 1, line);
 		const argument = args[0];
 		switch (argument) {
+			case NumIndex.index_0:
 			case json_refTypes.JSONString: {
 				this.jsonType = "byte";
 				break;
 			}
+			case NumIndex.index_1:
 			case json_refTypes.JSONUint64: {
 				this.jsonType = "int";
 				break;
 			}
+			case NumIndex.index_2:
 			case json_refTypes.JSONObject: {
 				this.jsonType = "object";
 				break;
@@ -5316,7 +5347,7 @@ export class Block extends Op {
 	constructor(args: string[], line: number, interpreter: Interpreter) {
 		super();
 		assertLen(args.length, 1, line);
-		const argument = args[0];
+		const argument = this.assertFieldIndex(args[0], blockFieldIndex);
 		if (argument === blockFieldTypes.BlkSeed || argument === blockFieldTypes.BlkTimestamp) {
 			this.field = argument;
 		} else {
@@ -5359,10 +5390,12 @@ export class VrfVerify extends Op {
 		super();
 		this.line = line;
 		switch (argument) {
+			case NumIndex.index_0:
 			case vrfVerifyFieldTypes.VrfAlgorand: {
 				this.vrfType = argument;
 				break;
 			}
+			case NumIndex.index_1:
 			case vrfVerifyFieldTypes.VrfStandard: {
 				this.vrfType = argument;
 				break;